Refactors into lib fns and folder
