+++ packages/jazz-richtext-prosemirror/src/lib/index.ts
@@ -1,0 +1,3 @@
+export { richTextToProsemirrorDoc } from "./document.js";
+export { applyTrToRichText } from "./transactions.js";
+export type { MarkType, NodeType } from "./types.js";
